Write usage examples, and test them in Travis
Every example is tested in Travis CI.  They all together comprise a test suite
for `gpg-build-scripts`.  Every example is arranged as follows: it begins
with a variable definitions block, which is followed with actual installation
commands, and ends with a sequence of assertions which confirm installation
correctness.  Every example includes comments.