Use own ld.so.conf file
Store paths to shared libraries in a dedicated file with more meaningful
name, that is "/etc/ld.so.conf.d/gpg-from_build_scripts.conf".

In the result, this file can be overwritten without serious risk that
it contains some important but unrelated paths, which should not be
erased.  In fact, file overwrites become idempotent, which makes
ldconfig paths management much easier.