Add --ldconfig option
Run ldconfig in post-install steps when --ldconfig option is given,
not when --sudo is given, as it was before.  Thanks to this change,
it is now possible to skip ldconfig while keeping --sudo option enabled.
This is especially important on systems which do not feature ldconfig
like OS X.  It also comes handy when user prefers to further customize
system libraries configuration, or to use $LD_LIBRARY_PATH instead,
for any reason.