Set Bash as programming language in Travis CI
This scritps are written in Bash, therefore it makes to specify Bash
programming language in Travis CI.  According to Travis documentation,
"Bash" is actually equivalent to "minimal".  However the former is more
meaningful and descriptive.  Also, if something changes, "Bash" will
more likely be the option we prefer.

See: https://docs.travis-ci.com/user/languages/minimal-and-generic/