Set project's programming language in Travis CI configuration (#43)
* Set Bash as programming language in Travis CI

This scritps are written in Bash, therefore it makes to specify Bash
programming language in Travis CI.  According to Travis documentation,
"Bash" is actually equivalent to "minimal".  However the former is more
meaningful and descriptive.  Also, if something changes, "Bash" will
more likely be the option we prefer.

See: https://docs.travis-ci.com/user/languages/minimal-and-generic/

* Remove empty build steps from Travis configuration

When language is set to "minimal" (or "bash", which is equivalent),
all build steps are empty by default.  There is no need to clear
something what is empty already.